+++ sitemap/twig.py
@@ -1,5 +1,7 @@
 from bs4 import BeautifulSoup
 import requests
+import os.path
+import sys
 
 # I switched around to the requests module on this one because I was told it yields better results.
 
@@ -15,7 +17,13 @@ dictionary = {}
 i = 0
 
 # Opens the output sitemap file.
-file = open("twig-sitemap.txt","w")
+if(os.path.isfile("twig-sitemap.txt")):
+    print("Sitemap already exists, either from a previous scrape, or user has generated a custom sitemap.\n")
+    print("Not generating sitemap.\n")
+    sys.exit(0)
+else:
+    file = open("twig-sitemap.txt","w")
+    print("File opened\n")
 
 # Finds every url in the sitemap that hold 2014/15/16/17/18 since that's the naming convention followed,
 # and adds it to a dictionary.
